Avoid errors in em_converged() due to NAs
It may happen that the value of the log-likelihood function is undefined and represented by NA. In that case, the comparison for convergence will error out as the missing value will propagate and break the if clause.

The primary reason for the undefined log-lik value is unclear. It is driven by a negative determinant of a (usually positive definite?) matrix in the SKF() function. Computing the log of the negative determinant will produce NAs.

The specific case where NAs are produced relates to GDP growth in 2009 Q1. The value is an outlier, hence convergence may be difficult due to unusual data properties. In effect, the loop runs to max_iter=500 (hard-coded in package).